fix: moved callback_lookup a un novo arquivo
